🔀 Implement memoised vdom nodes. (#405)
* :zap: make `map` its own node type

this improves performance by 3-10%.

* :sparkles: Introduce the public `memo` API

* :sparkles: Implement ref / dependencies

* :construction: add a memo node type and a ConcreteTree, pass things to
all the places.

This never materialises or diffs or stores lazy nodes, so it doesn't do
anything useful yet.

* :recycle: use a different type for partial diff results to clean up the
public API

* :construction: an extremely messy and untested but working version of
memo

* :zap: move subtrees to map nodes instead of lazy nodes.

this simplifies diffing a bunch since we don't have to track the mapper
anymore. It's a _required_ change for lazy to not have mappers inside a
single events tree.

It's also faster!

* :bug: fix map and is_controlled

* :recycle: rename map.element to map.child

* :sparkles: add fallbacks.

these are technically not necessary since it can :sparkles: never happen
:sparkles: but we should not write code that can crash, even in theory.

* :recycle: mutable_map.get_or_compute

* :recycle: remove direct tree/events interactions

* :recycle: ConcreteTree -> Cache, events.gleam -> cache.gleam

* :white_check_mark: add runtime smoke tests to hopefully catch syntax and
import errors in the future

* :memo: explain how memo/subtrees work a bit

* :white_check_mark: add memo tests

---------

Co-authored-by: yoshie <802632-arkandos@users.noreply.gitlab.com>